Moved the code for constructing the IQHolder query Arg to the service logic and added a new query in the service logic to handle specific start and end dates.
