+++ src/App/IQHolders/IQHolder.service.ts
@@ -1,18 +1,19 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
-import { Injectable, Inject, CACHE_MANAGER } from '@nestjs/common'
-import { Cache } from 'cache-manager'
-import { Cron, CronExpression, SchedulerRegistry } from '@nestjs/schedule'
-import { HttpService } from '@nestjs/axios'
-import { ConfigService } from '@nestjs/config'
-import { ethers } from 'ethers'
-import { DataSource } from 'typeorm'
-import IQHolderRepository from './IQHolder.repository'
-import IQHolder from '../../Database/Entities/iqHolder.entity'
-import IQHolderAddressRepository from './IQHolderAddress.repository'
-import erc20Abi from '../utils/erc20Abi'
-import IQHolderAddress from '../../Database/Entities/iqHolderAddress.entity'
-import { stopJob } from '../StakedIQ/stakedIQ.utils'
-import { firstLevelNodeProcess } from '../Treasury/treasury.dto'
+import { Injectable, Inject, CACHE_MANAGER } from '@nestjs/common';
+import { Cache } from 'cache-manager';
+import { Cron, CronExpression, SchedulerRegistry } from '@nestjs/schedule';
+import { HttpService } from '@nestjs/axios';
+import { ConfigService } from '@nestjs/config';
+import { ethers } from 'ethers';
+import { DataSource } from 'typeorm';
+import IQHolderRepository from './IQHolder.repository';
+import IQHolder from '../../Database/Entities/iqHolder.entity';
+import IQHolderAddressRepository from './IQHolderAddress.repository';
+import erc20Abi from '../utils/erc20Abi';
+import IQHolderAddress from '../../Database/Entities/iqHolderAddress.entity';
+import { stopJob } from '../StakedIQ/stakedIQ.utils';
+import { firstLevelNodeProcess } from '../Treasury/treasury.dto';
+import IQHolderArgs from './IQHolders.dto';
 
 export const IQContract = '0x579CEa1889991f68aCc35Ff5c3dd0621fF29b0C9'
 
@@ -36,6 +37,26 @@ class IQHolderService {
   private etherScanApiKey(): string {
     return this.configService.get<string>('etherScanApiKey') as string
   }
+
+async getIQHoldersWithArgs (args: IQHolderArgs): Promise<IQHolder[]> {
+  const iqHolderArgs: IQHolderArgs = {
+    interval: args.interval,
+    start: args.start,
+    end: args.end,
+    limit: args.limit,
+    offset: args.offset,
+  };
+  const iqHolders = await this.repo.getIQHoldersCount(iqHolderArgs);
+  return iqHolders;
+}
+
+async getIQHoldersByDateRange(startDate: Date, endDate: Date): Promise<IQHolder[]>{
+  const iqHolders = await this.repo.createQueryBuilder("iqHolder")
+  .where("iqHolder.updated >= :startDate", { startDate })
+  .andWhere("iqHolder.updated <= :endDate", {endDate })
+  .getMany();
+  return iqHolders;
+}
 
   async lastHolderRecord(): Promise<IQHolder[]> {
     return this.repo.find({
